Fixed warning.

+++ src/renderer_d3d12.cpp
@@ -4271,7 +4271,7 @@ data.NumQualityLevels = 0;
 
 			case Texture3D:
 				resourceDesc.Dimension        = D3D12_RESOURCE_DIMENSION_TEXTURE3D;
-				resourceDesc.DepthOrArraySize = m_depth;
+				resourceDesc.DepthOrArraySize = uint16_t(m_depth);
 				m_srvd.ViewDimension                 = D3D12_SRV_DIMENSION_TEXTURE3D;
 				m_srvd.Texture3D.MostDetailedMip     = 0;
 				m_srvd.Texture3D.MipLevels           = numMips;
